Lots of changes for Naomi's debugging for broken email rep forms
Changed aaron's email address to one Naomi can check.

Added code to handle special cases.  Much of this is documented in the README file or in the brokensenators() function.

Added a function senatetest2, and a new usage for command line:
 python main.py senatetest2 coons

So you can just enter in the name of the senator you want to email.