Changes to finish judiciary.
Every member of judiciary works except the two with captchas.

Added in a lookup table with addresses for the contact pages which actually do a street address check.
Made writerep function try harder.  If it fails with an exception when calling the writerep_ima or writerep_zipauth subroutines, it keeps trying.  The last resort is getting the hard-coded contact form.